# products/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  GenericError:
    description: GenericError is a generic error message returned by a server
    properties:
      message:
        type: string
        x-go-name: Message
    type: object
    x-go-package: build-microservice-with-go/products/utils
  Product:
    properties:
      created_at:
        type: string
        x-go-name: CreatedAt
      deleted_at:
        type: string
        x-go-name: DeletedAt
      description:
        description: |-
          the description for  this product.
          required: true.
        type: string
        x-go-name: Description
      id:
        description: the id for this user.
        format: int64
        minimum: 1
        type: integer
        x-go-name: ID
      name:
        description: |-
          the name for this product.
          required: true.
        type: string
        x-go-name: Name
      price:
        description: |-
          the price for this product.
          required: true.
          greater than: 0
        format: float
        type: number
        x-go-name: Price
      sku:
        description: |-
          the sku for this product.
          it should be a valid sku format.
          required: true.
        example: abcdasdas-abcea-abcd
        type: string
        x-go-name: SKU
      updated_at:
        type: string
        x-go-name: UpdatedAt
    required:
    - id
    title: Product type.
    type: object
    x-go-package: build-microservice-with-go/products/models
  ValidationError:
    description: ValidationError is a collection of validation error messages
    properties:
      messages:
        items:
          type: string
        type: array
        x-go-name: Messages
    type: object
    x-go-package: build-microservice-with-go/products/utils
info:
  description: Documentation for Product API
  title: of Product API
  version: 1.0.0
paths:
  /products:
    get:
      operationId: listProducts
      responses:
        "200":
          $ref: '#/responses/productsResponse'
      summary: Returns a list of products.
      tags:
      - products
    post:
      operationId: storeProduct
      responses:
        "200":
          $ref: '#/responses/productResponse'
      summary: store product.
      tags:
      - products
  /products/{id}:
    delete:
      description: Delete a products by id
      operationId: deleteProduct
      parameters:
      - description: The id of the product for which the operation relates
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "201":
          $ref: '#/responses/noContentResponse'
        "404":
          $ref: '#/responses/errorResponse'
        "501":
          $ref: '#/responses/errorResponse'
      tags:
      - products
    get:
      description: Return a list of products from the database
      operationId: listSingleProduct
      parameters:
      - description: The id of the product for which the operation relates
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/productResponse'
        "404":
          $ref: '#/responses/errorResponse'
      tags:
      - products
    put:
      operationId: updateProduct
      parameters:
      - description: |-
          Product data structure to Update or Create.
          Note: the id field is ignored by update and create operations
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/Product'
      - description: The id of the product for which the operation relates
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/productResponse'
      summary: update product.
      tags:
      - products
produces:
- application/json
responses:
  errorResponse:
    description: Generic error message returned as a string
    schema:
      $ref: '#/definitions/GenericError'
  errorValidation:
    description: Validation errors defined as an array of strings
    schema:
      $ref: '#/definitions/ValidationError'
  noContentResponse:
    description: No content is returned by this API endpoint
  productResponse:
    description: Data structure representing a single product
    schema:
      $ref: '#/definitions/Product'
  productsResponse:
    description: A list of products
    schema:
      items:
        $ref: '#/definitions/Product'
      type: array
schemes:
- http
swagger: "2.0"